requirements.txt
PyYAML
regex
PyPDF2